moved channel module into dma module
